# Store score in datalog file too