Bit of optimization for index rebuild and preparation to moving of index management into datalog. No bench penalty, only disk one

```
benchmark                old ns/op     new ns/op     delta
BenchmarkIndexLoad-4     78380828      78160618      -0.28%
BenchmarkOpen-4          77846975      77664061      -0.23%
BenchmarkWrite-4         23334         23379         +0.19%
BenchmarkRead-4          3983          4030          +1.18%

benchmark                old allocs     new allocs     delta
BenchmarkIndexLoad-4     9623           9608           -0.16%
BenchmarkOpen-4          9635           9641           +0.06%
BenchmarkWrite-4         1              1              +0.00%
BenchmarkRead-4          1              1              +0.00%

benchmark                old bytes     new bytes     delta
BenchmarkIndexLoad-4     41072605      41071168      -0.00%
BenchmarkOpen-4          41074290      41078896      +0.01%
BenchmarkWrite-4         1259          1260          +0.08%
BenchmarkRead-4          16            16            +0.00%

```